Reorganize project structure: move components and features to src directory
- Move all components from root to src/components/
- Move all features from root to src/feature/
- Move constants to src/constants/
- Update import paths in AppNavigator.tsx
- Maintain existing functionality while improving code organization
