+++ HealthTracker/src/components/CameraModal/index.tsx
