+++ HealthTracker/src/components/CameraModal/styles.ts
