+++ HealthTracker/src/components/LabeledInput/index.tsx
