+++ HealthTracker/src/components/LoadingModal/index.tsx
