+++ HealthTracker/src/components/LoadingModal/styles.ts
