+++ HealthTracker/src/navigation/AppNavigator.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
-import HomeScreen from "../../feature/HomeScreen";
+import HomeScreen from "../feature/HomeScreen";
 
 const Stack = createNativeStackNavigator();
